+++ Tests/KeyboardKitTests/Mocks/MockAutocompleteSuggestionProvider.swift
@@ -1,5 +1,5 @@
 //
-//  DisabledAutocompleteSuggestionProvider.swift
+//  MockAutocompleteSuggestionProvider.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-03-17.
@@ -7,20 +7,19 @@
 //
 
 import Foundation
+import KeyboardKit
 
-/**
- This internal class is used as a placeholder provider until
- a real provider is injected.
- */
-class DisabledAutocompleteSuggestionProvider: AutocompleteSuggestionProvider {
+class MockAutocompleteSuggestionProvider: AutocompleteSuggestionProvider {
     
     var locale: Locale = .current
     
+    var autocompleteSuggestionsResult = AutocompleteResult.success([])
+    
     func autocompleteSuggestions(for text: String, completion: (AutocompleteResult) -> Void) {
-        completion(.success([]))
+        completion(autocompleteSuggestionsResult)
     }
     
-    var canIgnoreWords: Bool { false }
+    var canIgnoreWords: Bool = false
     
     var ignoredWords: [String] = []
     
@@ -30,7 +29,7 @@ class DisabledAutocompleteSuggestionProvider: AutocompleteSuggestionProvider {
     
     func removeIgnoredWord(_ word: String) {}
     
-    var canLearnWords: Bool { false }
+    var canLearnWords: Bool = false
     
     func hasLearnedWord(_ word: String) -> Bool { false }